+++ Project/src/utilities/utils.hpp
@@ -12,9 +12,9 @@
 const float PI = 3.14159265359f;
 
 namespace UTILS
-{
+{ 
     glm::mat4 getViewMatrix(glm::vec3 position, glm::vec3 direction, glm::vec3 up = glm::vec3(0, 1, 0));
-    glm::mat4 getPerspectiveMatrix(int FOV);
+    glm::mat4 getPerspectiveMatrix(float FOV);
     unsigned int generateBuffer(Mesh &mesh);
     unsigned int initTexture(Image texture);
     void addTextureMaps(Appearance *appearance,
